V0.7: Fix export button breaking the form
Rather than trying to submit both format=raw and task=export in the export
button, simply use format=raw only and modify the controller to understand
format=raw as export.

The old way of using JS to set task=export was unable to reset the task and
thus broke all further use of the form controls, because then task=export was
submitted, even though no task was supposed to be used.

Fixes #3